requirements.txt
streamlit
spacy
nltk
scikit-learn
python-docx
pdfplumber
fpdf
langdetect
